+++ src/app/features/cart/cart.module.ts
@@ -1,6 +1,8 @@
 import { NgModule } from '@angular/core';
 import { CommonModule } from '@angular/common';
 import { CartListComponent } from './components/cart-list/cart-list.component';
+import { CartItemComponent } from './components/cart-item/cart-item.component';
+import { FontAwesomeModule } from '@fortawesome/angular-fontawesome';
 
 const EXPORTED_COMPONENTS = [
   CartListComponent
@@ -8,10 +10,12 @@ const EXPORTED_COMPONENTS = [
 
 @NgModule({
   declarations: [
-    ...EXPORTED_COMPONENTS
+    ...EXPORTED_COMPONENTS,
+    CartItemComponent
   ],
   imports: [
-    CommonModule
+    CommonModule,
+    FontAwesomeModule
   ],
   exports: [
     ...EXPORTED_COMPONENTS