+++ src/app/features/cart/models/cart-item.ts
@@ -3,10 +3,10 @@ import { ProductModel } from '../../products/models/product';
 
 export class CartItem {
   constructor(
-    public item: ProductModel,
+    public product: ProductModel,
     public amount: number) { }
 
   getTotalItemPrice(): number {
-    return this.amount * this.item.price;
+    return this.amount * this.product.price;
   };
 }