+++ src/app/shp.module.ts
@@ -1,5 +1,9 @@
 import { NgModule } from '@angular/core';
 import { BrowserModule } from '@angular/platform-browser';
+import { FaIconLibrary, FontAwesomeModule } from '@fortawesome/angular-fontawesome';
+import { fas } from '@fortawesome/free-solid-svg-icons';
+import { fab } from '@fortawesome/free-brands-svg-icons';
+import { far } from '@fortawesome/free-regular-svg-icons';
 
 import { ShpRoutingModule } from './shp-routing.module';
 import { ShpComponent } from './shp.component';
@@ -7,7 +11,6 @@ import { ProductsModule } from './features/products/products.module';
 import { CartModule } from './features/cart/cart.module';
 import { SharedModule } from './shared/shared.module';
 import { OrdersModule } from './features/orders/orders.module';
-
 @NgModule({
   declarations: [
     ShpComponent,
@@ -19,9 +22,14 @@ import { OrdersModule } from './features/orders/orders.module';
     SharedModule,
 
     BrowserModule,
-    ShpRoutingModule
+    ShpRoutingModule,
+    FontAwesomeModule,
   ],
   providers: [],
   bootstrap: [ShpComponent]
 })
-export class AppModule { }
+export class AppModule {
+  constructor(library: FaIconLibrary) {
+    library.addIconPacks(fas, far, fab);
+  }
+}