+++ changelog.md
@@ -12,6 +12,10 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ### Added
 - ProductsModule, CartModule, OrdersModule, SharedModule
 - CartListComponent displays total price
+- FontAwesome
+- CartItemComponent
+- Ability to edit cart amount
+- Ability to remove amount
 
 ## [0.1.1] - 2021-09-15
 ### Changed